pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["ros_wheel_builder"]

[project]
name = "ros-wheel-builder"
version = "0.0.1"
dependencies = [
    "rosdistro",
    "jinja2",
    "jupyterlab",
    "requests",
    "pyyaml",
    "rosdep",
    "lxml",
    "toposort",
    "cibuildwheel",
    "ruff",
    "fire",
    "twine"
]

[project.scripts]
build-ros-wheel = "ros_wheel_builder.main:main"
upload-ros-wheel = "ros_wheel_builder.upload:main"